Second update since presentation
For this round of updates, I focused on the DELETE and cleaning up the
readability of my code.  I managed to get DELETE for the animal resource
running, and I went through to add additional comments/move things
around to make the code a bit easier to read.  At this point, the
functionality is 100% and now I can use the extra time to focus on
the CSS and styling aspects.