3rd update since presentation
This round of updates was entirely visual/CSS focused.  I removed a
large portion of CSS styling I had previously input in an effort to try
and implement more Bootstrap.  I also discovered Bootstrap has a section
of icons that you can utilize in your HTML, so I added a few of those in
to some of my various buttons as well.  As a next step, I'd like to
clean up my layout of my different components on the page.