CSS/Bootstrap UI update
For this round of updates, I cleaned up the readability of my jQuery
and the success of messages generating/removing at the correct times.  I
also modified the CSS to align my elements more towards the center.  As
a next step, I'd like to resolve the issue where my SignUp/signIn
collapse over each other when the page size has shrunk - it's not
impacting any of my other elements so it will take a bit of time to
review the cause of the issue.