Final Update for Submittal
As a final update, I was able to resolve my Update issue - it turns
out I had snapped the connection via my HTML so the Mongoose connection
was unable to read it.  Also, as requested, I updated my README file to
contain the links to the alternate repo and to the deployed site and
the API.  I slightly moved around the UI/CSS for the footer and I
confirmed that each resource can CRUD properly in the deployed version.